+++ libs/dot-components/src/lib/components/dialog/Dialog.stories.tsx
@@ -1,13 +1,18 @@
 import React from 'react';
 import { Story, Meta } from '@storybook/react/types-6-0';
-
 import { DotDialog, DialogProps } from './Dialog';
 
 export default {
-  title: 'Experimental/Dialog',
+  title: 'Components/Dialog',
   component: DotDialog,
   argTypes: {
     children: { defaultValue: 'Batman is way better than superman' },
+    onCancel: {
+      action: 'cancelled',
+    },
+    onSubmit: {
+      action: 'submitted',
+    },
     open: { defaultValue: true },
     title: { defaultValue: 'Superheros' },
   },